refactor move EditorContent to layout/ directory
